requirements.txt
agent-framework-azure-ai>=0.0.1rc1
openai>=1.3.0
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
cryptography>=41.0.0
reedsolo>=1.7.0
numpy>=1.24.0
scipy>=1.11.0
